requirements.txt
openai
pandas
numpy
requests
python-dotenv
scikit-learn
chromadb
